pyproject.toml
[tool.poetry]
name = "neoscore"
version = "0.1.2"
description = "A graphical musical notation library"
authors = ["Andrew Yoon <andrew@nothing-to-say.org>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://neoscore.org"
repository = "https://github.com/DigiScore/neoscore"
classifiers = ["Topic :: Multimedia :: Graphics"]
packages = [
    { include = "neoscore" }
]
[tool.poetry.urls]
"Bug Tracker" = "https://github.com/DigiScore/neoscore/issues"

[tool.poetry.dependencies]
python = "^3.7"
PyQt5 = "^5.15.6"
img2pdf = "0.4.4"
sortedcontainers = "2.4.0"
typing_extensions = "^4.3"
"backports.cached-property" = "1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
pytest-xdist = "^2.5"
Sphinx="^4.5"
sphinx-rtd-theme = "1.0.0"
beautifulsoup4 = "^4.11.1"
lxml = "^4.8.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

